Move all interfaces to the Internal namespace
