+++ src/Microsoft.AspNetCore.ResponseCaching/Internal/ResponseCacheContext.cs
@@ -6,10 +6,9 @@ using System.IO;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Http.Features;
 using Microsoft.AspNetCore.Http.Headers;
-using Microsoft.AspNetCore.ResponseCaching.Internal;
 using Microsoft.Net.Http.Headers;
 
-namespace Microsoft.AspNetCore.ResponseCaching
+namespace Microsoft.AspNetCore.ResponseCaching.Internal
 {
     public class ResponseCacheContext
     {
