+++ src/Microsoft.AspNetCore.ResponseCaching/Internal/ResponseCachePolicyProvider.cs
@@ -6,7 +6,7 @@ using Microsoft.AspNetCore.Http;
 using Microsoft.Extensions.Primitives;
 using Microsoft.Net.Http.Headers;
 
-namespace Microsoft.AspNetCore.ResponseCaching
+namespace Microsoft.AspNetCore.ResponseCaching.Internal
 {
     public class ResponseCachePolicyProvider : IResponseCachePolicyProvider
     {
